- Upgrade Sequel Pro to be compiled as a 3-way PPC/i386/x86_64 binary for release builds, including a large number of 64bit compatibility upgrades and tweaks  - Upgrade RegexKitLite to 3.3

+++ Source/CMCopyTable.h
@@ -70,7 +70,7 @@
 	@param	  isLocal who cares
 	@result	 Always calls for a copy type drag operation
 */
-- (unsigned int)draggingSourceOperationMaskForLocal:(BOOL)isLocal;
+- (NSUInteger)draggingSourceOperationMaskForLocal:(BOOL)isLocal;
 
 /*!
 	@method	 selectedRowsAsTabStringWithHeaders
@@ -107,5 +107,5 @@
 
 @end
 
-extern int MENU_EDIT_COPY_WITH_COLUMN;
-extern int MENU_EDIT_COPY_AS_SQL;
+extern NSInteger MENU_EDIT_COPY_WITH_COLUMN;
+extern NSInteger MENU_EDIT_COPY_AS_SQL;
