+++ Source/SPSSHTunnel.h
@@ -62,23 +62,23 @@ enum spsshtunnel_password_modes
 	BOOL useHostFallback;
 	BOOL requestedResponse;
 	BOOL passwordInKeychain;
-	int sshPort;
-	int remotePort;
-	int localPort;
-	int localPortFallback;
-	int connectionState;
+	NSInteger sshPort;
+	NSInteger remotePort;
+	NSInteger localPort;
+	NSInteger localPortFallback;
+	NSInteger connectionState;
 }
 
-- (id) initToHost:(NSString *) theHost port:(int) thePort login:(NSString *) theLogin tunnellingToPort:(int) targetPort onHost:(NSString *) targetHost;
+- (id) initToHost:(NSString *) theHost port:(NSInteger) thePort login:(NSString *) theLogin tunnellingToPort:(NSInteger) targetPort onHost:(NSString *) targetHost;
 - (BOOL) setConnectionStateChangeSelector:(SEL)theStateChangeSelector delegate:(id)theDelegate;
 - (void) setParentWindow:(NSWindow *)theWindow;
 - (BOOL) setPasswordKeychainName:(NSString *)theName account:(NSString *)theAccount;
 - (BOOL) setPassword:(NSString *)thePassword;
-- (int) state;
+- (NSInteger) state;
 - (NSString *) lastError;
 - (NSString *) debugMessages;
-- (int) localPort;
-- (int) localPortFallback;
+- (NSInteger) localPort;
+- (NSInteger) localPortFallback;
 - (void) connect;
 - (void) launchTask:(id) dummy;
 - (void) disconnect;
